Move components to separate directory
