+++ v3/src/js/views/components/cors-stats/CorsStats.jsx
