+++ v3/src/js/views/components/cors-stats/styles.scss
@@ -20,7 +20,7 @@
 }
 
 .semester {
-  margin-bottom: 4rem;
+  margin-bottom: 2rem;
 
   .semHeading {
     margin: 1rem 0 0.4rem;
@@ -69,7 +69,7 @@
   font-size: 1rem;
 
   svg {
-    width: 1rem;
+    width: 1.2rem;
     vertical-align: middle;
   }
 }
